# Conflicts:
#	dartagnan/src/main/java/com/dat3m/dartagnan/parsers/program/visitors/boogie/VisitorBoogie.java
#	dartagnan/src/main/java/com/dat3m/dartagnan/program/event/Tag.java
#	dartagnan/src/main/java/com/dat3m/dartagnan/program/event/core/Event.java
#	dartagnan/src/main/java/com/dat3m/dartagnan/program/processing/DeadAssignmentElimination.java
#	dartagnan/src/main/java/com/dat3m/dartagnan/program/processing/DeadCodeElimination.java
#	dartagnan/src/main/java/com/dat3m/dartagnan/program/processing/FindSpinLoops.java
#	dartagnan/src/main/java/com/dat3m/dartagnan/program/processing/LoopUnrolling.java
#	dartagnan/src/main/java/com/dat3m/dartagnan/program/processing/RemoveDeadCondJumps.java

+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/processing/DeadCodeElimination.java
@@ -60,7 +60,7 @@ public class DeadCodeElimination implements ProgramProcessor {
         Event cur = entry;
         int id = startId;
         while (cur != null) {
-            if (!reachableEvents.contains(cur) && cur != exit) {
+            if (!reachableEvents.contains(cur) && cur != exit && !cur.is(Tag.NOOPT)) {
                 cur.delete();
                 cur = pred;
             } else {
